requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
qdrant-client==1.7.0
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
InstructorEmbedding==1.0.1
python-dotenv==1.0.0
numpy==1.24.3
pydantic==2.5.0
aiofiles==23.2.1
pypdf2==3.0.1
python-docx==1.1.0 